Mini map
Created a working mini map. The script allows for a camera that follows the player. Using a render texture that is then put on that camera and then shows it on a raw image on the UI. There is a Mini map render texture in the assets folder, labeled as such.