Rotation Bug Fix
Fixed a bug which caused the X and Z rotations to be set to zero when turning, meaning going up any sort of ramp was clunky. This may cause an issue in the future if there is a surface that causes the player to topple over, however the framework is in place to fix it should it become an issue.