Added Coin Particles, Animations and more
I've now added particles to the coins on idle and on pick up. I've also made the score text bounce and change colour depending on score increase and decrease. I added a red vignette on collisions and I have replaced the player placeholder model with the new cart model. The model, however, is buggy due to raycasts needing changing.